Start work on Light effect
This operates like classic where a small candle appears in front of player and flits about. Light source moves with candle and casts shadows.
Still have a lot more to do with this. At this time magic candle object does not expire with spell effect and other target types are not supported.