Merging FLC code by @lypyl
Lypyl PM'd this to me way back in Aug 2015. Somehow it never made it into the code. Correcting that now.
This is a reader for .flc files such as AZURA.FLC, the animations played when player meets various Daedric Princes.
Just merging code as-is for now. It can be reviewed and wired up to game flow when Daedric summoning is ready to be implemented.