Progress Light effect
Renamed Light class to LightNormal just to avoid any confusion with Unity's internal Light class. This doesn't change any of the information displayed to player or how effect is keyed, just the class name.
Set MagicCandle to expire when player changes game or effect ends normally.
Changed effect to be caster only for now. I intend to do more with Light in future, and this is the core effect purpose implemented now.