Implement new slope-friendly gravity system
Migrated the anti-bump factor from the FrictionMotor to the GravityMotor and implemented a raycast system that checks immediately in front of the player's feet to see if there's a downward slope. and uses higher rate of negative y movement when slopes or stairs are there.  and regular gravity when not.