Fix Short Rappeling bug, and Gravity Fix with bump factor
Creates a small component class for playerAdvanced called PlayerStepDetector that detects what the player is about to step onto by measuring the distance detected by SphereCast downward, the origin of this ray is offset by the player Movement vector so it detects the distance to the floor of what the player is stepping onto.  The classis for efficiency and code reuse.

ClimbingMotor uses this to prevent rappeling from really short distances.  AcrobatMotor uses this to determine whether it should add -AntiBumpFactor to gravity or not.