Implement Free Action
Target becomes immune to effects of paralyzation while Free Action is running, but not immune to receiving a bundle with paralyzation as a payload.
If a paralze effect is running when Free Action expires, then incumbent paralyzation will become active again.
This is creates a clear use-case difference between Free Action and Cure Paralyzation (which actually removed paralyze incumbents).