Fix non-elemental effects using wrong element for saving throws
It's possible to combine multiple effects in a single spell. While the spellmaker UI lets you select an element for the spell itself, component effects might be a mixture of elemental and magic-only.
GetElementType() now returns Magic for magic-only effects. All other effects will use the element selected for the parent spell bundle.
Now if a magic-only effect (e.g. Drain Strength) is combined with an elemental effect (e.g. Damage Health) and an element like Fire is selected, then Fire will be selected for elemental effect and Magic will be selected for magic-only effect for purposes of saving throws.