Destroy invalid QuestResourceBehaviour on load
When loading a save where enemy is tagged as a quest spawn, but not linked to an actual quest, the QuestResourceBehaviour should be destroyed. Otherwise enemy may become immortal as pinned to a non-existent quest.
Still looking for root cause of QuestResourceBehaviour being improperly added in first place, but this will fix issue when loading an affected save.