fix all known shortcomings of stack splitting
CanCarry() have been replaced by CanCarryAmount() that return the
quantity of item that can be carried or hold; It's used to know if
stacks could be partially carried.

SplitStack returns the original stack if all items are picked, that
should no longer be a special case.

Since TransferItem() can be asynchronous, I need to pass along another
flag to know if an item equipping should be attempted after transfer.
Not too fond of that.

And arrows have a weight now.