Store gender of static NPC Questor
This copies gender from quest giver in world, and does not allow that to be overwritten by definition in quest. This small change should solve a number of problems with incorrect face and name compared to NPC just clicked on.
Some other gender switching might remain from different causes, such as a poorly written quest script.