Enables player to switch between hanging and climbing
the switch from hanging to climbing is a bit quirky at this point though, It checks twice, and creates two climbing messages.  and gravity from the transition accumulates each time it is done, making the player grab the wall lower and lower each time it is done until gravity is reset.