Handle null entityBehaviour.Target
entityBehaviour.Target is checked for null, but no action is taken when it is null and can be subsequently used even if null. Under some conditions can cause an exception.
Added an else statement to exit when enityBehaviour.Target is null so following code does not try to access it.