ImportedComponentAttribute: Check unhandled component references
During deserialization, a component instance must be created with GameObject.AddComponent() instead of the constructor. This is handled by our custom converter, but if a reference to a component is stored in a field in editor (as opposed to be retrieved with GetComponent()), the serialized data will be located on the first reference met.

If the first reference is one of these fields (and not the actual component on the gameobject), deserialization will fail. To avoid run time issues, serialization now detects this situation and is failed with a message.