Allow HUD render toggle with LeftShift+F10
Toggles HUD with keypress.

Current list of reserved hard-bound keys in DFU:
- Tilde/Backquote: Console
- LeftShift+Tab: Quest debugger
- LeftShift+F10: Toggle HUD
- LeftShift+F11: Toggle SDF font rendering

It's really getting to the point where a custom binding interface is needed for these special keys. Particularly the console key which can be hard for end users to toggle with some non-US keyboard layouts.

Not putting priority on this right now, as it's ideally requires a whole new UI. Will also need to decide if this replaces the classic binding UI completely or works in tandem with it.