only select start marker based on incoming direction when travelling
Player can be relocated "to a random marker" in 6 cases:
- when entering/exiting buildings and dungeons;
- after fast travel;
- when going from/to ship;
- when teleported;
- when expelled by court;
- using "location" console command.

This patch implements start coordinates saving, and selection of facing
based on those coordinates as a new RelocationMethod; And then use this
method for fast travel case only.

There's some comment near entering/leaving buildings
(PlayerEnterExit.Respawner) that suggests that the closest start marker
should be chosen, but it's unclear if this new RelocationMethod could
help.