Delay inventory scroll up when removing items
When removing an item from inventory, bottom items are moved up; Unless
the end of the list is already displayed, in which case the inventory is
also scrolled up, so it looks like it's the top items that are moved
down.
It gets worse with the two columns of the enhanced inventory, because if
the end of the list is displayed, removed item will be alternatively
replaced by the item to its right and by the items above it, depending
on whether last row was full or only half filled.

To keep it short, scrolling while removing items is confusing, let's
avoid that!