Setting version to 0.7.0
We've been in a holding pattern making passes over 0.5 and 0.6 combined for some time now. I'm happy with the current 0.5 stable build and there's not much in 0.6 that hasn't already been in the game for some time now. I should really have called the latest build 0.6 stable.
I'm skipping straight to 0.7 now and will start working on new features in that queue soon.