Prepare for racial override effect
Vampire, Werewolf, and Wereboar are no longer races in their own right. This will be done with racial overriding from effect system. More to come on this soon.
Commented out GetVampireOrWereCreatureQuest() as this might be deployed from a different system down the road.
When importing a vampire or lycanthrope their secondary race is stripped away. Vampire bonuses are also removed on import. Lycanthropes are todo.
Once racial overriding is functional the appropriate effect for transformed race needs to be instantiated again on player. All lost bonuses will be restored by effect.