Cache racial override effect in effect manager
Player entity was using FineIncumbentEffect<RacialOverrideEffect>() - a brute force search for a particular effect. This is usually very quick as entities typically have a very low number of active effects (usually < 10).
But player race is used in some higher-frequency situations like combat formulas, and I want this to be as fast as possible.
Created a specialised reference cache for racial overrides so new race will be provided as quickly as possible for systems calling PlayerEntity.RaceTemplate regardless of how many effects player has running.
EntityEffectManager will cache RacialOverrideEffect reference on start/resume and clear it when removing its bundle or wiping all effects.