HUD place markers now support floating origin
When travelling or loading directly to within a city, place markers might end up not positioned properly. This could be solved by running out of city limits and back in again, but that was rather inconvenient.
Added a new event for floating origin changes. HUD place markers now listen for this event and will include floating origin in marker position calculation.