FACTION.TXT is now part of StreamingAssets
This text file is notorious for breaking game compatibility based on patch level, where the game was installed from, target file system, and the presence of optional translation packs. This single file is responsible for more compatibility problems than any other.
A known-good and nicely patched version of FACTION.TXT is now included with game in StreamingAssets/Factions folder. Where this file was file was previously being read from Arena2 folder, it will now be read from StreamingAssets.
End users are still free to replace/edit this as they wish.