Add lower range check
A user sent me a save where socialGroup was a negative number, resulting in an out of range exception for
reactionToPlayer += player.SGroupReputations[(int)socialGroup];
Added a check just to ensure socialGroup is >= 0. This is probably indicative of some of other data problem, but checking range before accessing an array is good practice in any case.
In this case, NPC was Medora in Direnni. I will attach save in comments.