Improve handling of MessageBox with buttons
It was possible to close a MessageBox with buttons without providing required input either through cancelling (if owner forgot to set AllowCancel=false) or using return/enter/custom binding. A MessageBox with buttons is expecting some clear input from user and it should not be trivial to skip this.
A MessageBox will now automatically set AllowCancel=false when a button is added.
Buttons added to a MessageBox can now be set as the default button. Only one button can carry this flag and any other buttons set as default will be unset if a new default button is added to collection.
When using enter/return/custom to close a MessageBox, the default button  (if one is defined) will be manually triggered.
If no default button present, and another MessageBox is not queued, then enter/return/custom action is ignored so that player must provide input requested by MessageBox owner.
The common "yes"/"no" MessageBox button format now defaults to "no" button.