pay for spellmaker, check for spellbook presence
Supersedes #1252 (Deducts gold when buying spells from spellmaker)

Also, check for spellbook presence. In classic, when you have no
spellbook:
- spell merchant does nothing, not explaining why;
- spell maker fails at the last minute ("you have no spellbook to
inscribe this spell")

I think that in both cases the player, even without a spellbook, could
want to peruse available spells, so I implemented a preliminary warning
that does not block the access to the spell merchant and the spell
maker.
A second check is done if you try to buy a spell though.

Maybe that's overkill and confusing, and access could simply be denied
with an error message?