Wiring to execute enchantment payloads
Enchantment param is stored in a flexible EnchantmentParam struct. Currently this serves legacy and custom enchants, and can be expanded later if required.
Enchantment params are passed to live instances of enchantment effect bundles or using callback method.
Callbacks are executed directly on template for stateless payloads. If an enchantment needs to be long-running with a full effect-entity relationship, it should specify the Held flag. Enchantments with a Held payload are assigned a live effect bundle on target entity for the duration enchanted item is equipped.
Currently EntityEffectManager still executing "when held" spell enchantments using the old setup. This will be migrated to new enchantment class.
PassiveItemSpecialsEffect is now deprecated and will be removed once all payloads moved to their respective enchantment class.