Enchantment exclusive test support
Some enchantments are exclusive to others and cannot be combined. This can happen on primary or secondary types, between side-effects and powers, or even with self.
IsEnchantmentExclusiveTo() is called for each effect when populating enchantment lists. The effect itself can determine if it's exclusive with any other effect in list or not. This design seems fairly simple and flexible.
Imeplemented exclusive tests in GoodRepWith, BadRepWith, FeatherWeight, and ExtraWeight.