Helpers to check if item contains enchantment
More robust than assuming a specific slot is populated as expected or that enchantments are even present.
Azura's Star item checks now use these helpers.
Would like to move more of Azura's Star to effect, but as soul trapping is built into framework this would likely require a larger refactor. Happy with Azura's Star as it is right now.