Change #350 faction type
This was previously 7 (province/region) which conflicts with Ilessan Hills. It seems more appropriate for this to be type 2 (group).
Corrects shared province resolution with Ilessan Hills and should not affect anything else.
Existing saves will still have #350 cached as type 7 so the previous fix of logging warning only is still required.