Initial import of classic lycanthrope character
At this time, correct lycanthropy type can only be determined if character is actively transformed at time of save. Unknown where lycanthropy type is stored while player is not transformed at time of save.
However, its possible to determine character is a lycanthrope of some kind as the "!Lycanthropy" spell record (ID=92) is present as a child of old class record data. Using this to alert player at import time that character should be transformed or lycanthropy is not restored. This will serve as a workaround for now until a better method of always identifying lycanthropy type is found.
As with vampirism, embedded bonuses are stripped away and racial override effect added to player (todo). This effect will add the bonuses back again, along with other payloads and management.