implement smarter location finder using edit distance
EditDistance contains a quite flexible edit distance function:

public EditDistance(Func<string, string> canonize_string,
                    Func<char, float> insert_cost,
                    Func<char, float> delete_cost,
                    Func<char, char, float> replace_cost,
                    Func<string, int, float> seek_cost,
                    Func<string, int, float> trim_cost)

- canonize_string(s) is the function that will be used to preprocess
dictionary and search strings
- insert_cost(c) is the cost function for adding a specific character
- delete_cost(c) is the cost function for removing a specific character
- replace_cost(c1, c2) is the cost function for replacing a specific
character with another
- seek_cost(s, start) is the cost function for seeking in the dictionary
string s to the start-th character
- trim_cost(s, stop) is the cost function for seeking in the dictionary
string s from the stop-th character to the end

Lowering the latter 2 costs (compared to insert_cost) allows to
approximate a substring search.

float distance(string s1, string s2, float upperBound =
float.PositiveInfinity)
This computes the edit distance to transform string s1 into string s2.
If an upperBound is provided, the function short-circuits and returns
float.PositiveInfinity if it detects during computation that the final
cost will be above upperBound. This requires all costs to be positive.

This class has a SetDictionary(List<String>) method to set the
dictionary that will be used to next searches. It canonizes all the
dictionary strings, so it could be useful to only call it when the
dictionary changes.

MatchResult[] FindBestMatches(string needle, int ntop)
Finds at most ntop dictionary entries that have the smallest distance to
needle string.
MatchResult is a structure that contains the (non-canonized) matching,
string, its distance, and a relevance. The results with the best
relevance (hence smallest distances) are returned at the beginning of
the list.

DaggerfallEditDistance is a static class that provides an EditDistance
object pre-configured with useful parameters:
- trimmed/lowercased strings canonization;
- smaller cost for adding or removing separators (space and dash)
compared to letters;
- smaller cost for replacing a separator with another;
- smaller cost for substring prefix search, and to a lesser extend
beginning of word search;
- smaller cost for substring suffix search, and to a lesser extend
ending of word search.

DaggerfallTravelMapWindow has been modified to use findBestMatches() of
this object; In case of approximate matches, a location picker window
will appear with name suggestions.
DaggerfallListPickerWindow was slightly modified to take keyboard focus
in this case.