if perfect match exist, return all perfect matches only
Normally there should be only one perfect match (distance == 0), but if
string canonization generates collisions that's no longer guaranteed.

Say, you set canonization to a function that strip strings of the prefix
"the ", as I think classic Daggerfall does, and you have both "X" and
"the X" in your dictionary.
Both will be canonized as "X", and whether you search for "X" or "the
X", both will match, and both will be perfect matches. In such case,
keep both in result, as they have different display strings.

(that may not be a very good example of canonization that produces
collisions, but hopefully it gets the idea across).