Retro 320x200 world rendering
When enabled, renders world to a 320x200 render texture that is blitted to an otherwise empty viewport camera.
Render texture is set to point filtering so it is presented with crisp chunky pixels and no blur.
UI remains at set screen resolution so it's possible to have a ultra-HD UI overlayed on a 320x200 world.
Most mods will still work, but other stacked camera mods like distant terrain are likely to break with this option enabled. People who use this setting are unlikely to use these mods anyway, and support can always be added to mods later.
Changes are only made when Retro320x200World is enabled. Otherwise rendering is exactly as before.