Fix night-day ambient transition
Night time ambient light is considered the lowest level of exterior ambient light. Because optional scale was only applied at night there was a sudden ramp-up in ambient intensity when the scaling factor was removed.
As this is the baseline for lerp, night time scale must be applied always to ensure smooth transition from night to day.