Fix macro %g3 and %pg3 to his/her instead his/hers
This macro is used in cases like "needs to be paid for %g3 item", which previously resulted in pronoun selection "needs to be paid for HERS item".
Have noted this in many classic quests. In some cases have changed to %g2, but this is clearly an emulation issue with macro expansion. Fixing macro itself now.