Place support for building name in dungeons
For example, when a noble quest is started inside Castle Daggerfall, the string "__questgiver_, ___questgiver_" should exapand to "Castle Daggerfall, Daggerfall". Previously building name was not being set for dungeons, which is where majority of nobles are found.
Added support for dungeon name expansion of this macro, with a fallback to town/location name if not inside a special dungeon.