Shadow distance settings
The universtal shadow distance property in QualitySettings was much too high for interiors and particularly dungeons. This resulted in CPU-bound framerates for most systems in dungeons with lots of lights.
Added settings to override shadow distance based on context. Adjusted shadow distance for each context to more reasonable defaults.
This will make dungeon shadows viable on a much wider range of systems.