Illegal rest warning option
When false, behaviour matches classic.
When true, player sees prompt "It is illegal to camp in or near a city. Continue?" and they can choose to continue with rest or not. They may even get in a few hours of actual vagrancy before the same inevitible outcome occurs.
This option adds both a safety net for accidental camping and makes vagrancy an option. Player will actually camp for a while before the watch arrives. I feel this also adds a bit more RPG flavour to resting inside towns.