only allow access to wagon when in sight
Patch on top of #1341

There's probably still some bug, but I need to find a dungeon to test
it:
Some dungeons have more than one exit door, but
DungeonWagonAccessProximityCheck() checks the proximity to the
StartMarker, not from a dungeon exit. If my hypothesis is correct, when
DungeonExitWagonPrompt is disabled you cannot use those secondary exit
doors.