Fixed issue with terrain metallicgloss maps
Metallicgloss maps import for terrain allows to provide only some
of them and make a default texture for missing records.

This default texture can only be created with RGBA32 or ARGB32 format, so it can't
be added to texture array with CopyTexture if array itself is using a different
format, like DXT.

Fallback Get/SetPixels copy is used instead for this situation. CopyTexture is still
used on supported systems when all records are required (meaning albedo and normals maps).