Lycanthropy should be a free spell
Even though classic displays a spellpoint cost in spellbook, casting Lycanthropy does not deduct spellpoints and player can always return to humanoid form even if they have 0 spellpoints.
This fixes issue so that lycanthropes cast for free, and spell displays correctly in DFU as 0 cost and always castable.
The once per day logic is not affected by this change.