FindClosestDoorToPlayer() to return false if all doors are filtered out
DaggerfallStaticDoors.FindClosestDoorToPlayer() was returning true
if all doors were filtered out by requiredDoorType.
When this happened, doorPosOut would still have its default value of
playerPos, making DaggerfallInventoryWindow believe that indeed a door
was found close to the player.

This only happens in some dungeons. The issue was first seen in a
stream, in the Citadel of Hawking, Daggerfall:

    https://www.twitch.tv/videos/461269349?t=01h50m27s