Quest dispense minor fixes
Changed "please wait" message to a simple click anywhere to dismiss popup. No need for user to aim at a button.
@ajrb There appears to be an issue with single-button height adjustments in UI system, which is why you had trouble. I'll look at this while I'm reviewing text. But we don't need a button here in any case.
Trimmed message slightly and delivered via tokens so that it can be centre justified.
Fixed name of C0B00Y02 to "The Possessed Child". This quest is about a possessed child, not an obsessed one.