Evaluate travel-to Place resources directly from message
PR #1426 made a change required for %di to store NPC home Place as last place resource referenced for talk system. While this is required for %di, it causes issues in journal for "travel to location" on click in two ways:
1. If journal Message does not expand a location then it will fallback to last NPC home place, which can be unrelated to quest. An example of this is DB initiation quest that mentions no actual locations in journal message, just a "I dare not copy this to my journal" message.
2. If a quest Person is mentioned after a quest Place, this will change LastPlaceReferenced to that of Person rather than Place.
Problem is related to how journal is determining which Place to select.
What this change does is directly enumerate the actual Place resources listed in quest log message just clicked on. Then uses the last Place resource mentioned in Message. If no Place is mentioned in Message, it will return null.
Behaviour is almost identical to previous implementation but will only use Place resources not NPC homes.
Ideally the quest journal travel-to click would hyperlink Place names so player can select which Place they wish to travel to when multiple are listed (e.g. gold bar recovery quest).
@ajrb Just tagging you so you know change has been made. I'm happy with outcome here, but feedback is welcome. Cheers.