Reworked and improved support for custom books
Book mapping manifest files must provide an unique id for each book,
chosen among all numbers representable with an integer.
This ensures that serialized inventory books are always resolved to the
correct book resource, even if save is loaded with different mods setup.

It's also possible to specify a questing global variable that must be set
before a book is available, or even set the IsUnique flag to prevent a book
to appear inside random loots.