Progress travel map refactor
Region textures are now created only once on first access and cached for lifetime of game. This does use a little texture memory over time, but is nothing compared to leaking gigabytes of texture memory like before.
Overlay pixel buffer and texture for region identify and crosshair are created only once and updated as needed.
Setup new identify and crosshair logic that only requires switching panel off and on. It does not need to recreate textures every frame during animation.
Still have a few more errors to fix and a cleanup.