Code cleanup after refactor
Travel map refactor should now match all prior behaviours.
Tested with classic and with DREAM for updated maps OK.
Need to test and look for any regressions. So far seems stables and functional.
Excessive texture creation and memory leaks have been addressed with new design.
Removed TempObject as this was not required in the end.