Implement Foe item queues
When using "give item" to assign an item to a Foe resource, an clone of this item will now be added to each subsequent instance of that Foe. I've always thought this was a wonky mechanic in classic, but some quests actually rely on this to move Foes to different locations while retaining assigned items.
This also solves the issue of a quest item only being placed on on an enemy (e.g. when player enters dungeon) and losing it if player resets dungeon by leaving and entering again.
@jayh2971 Jay, could I please ask for some help testing this one? Could you run it through some quests and let me know if any problems or unexpected behaviours. This is a fairly complex change to item placement, but I think I have all the serialization and condition angles covered so that it should be stable. Thank you for your help!