Support live mod settings changes
Adds support for live changes to mod settings. Mods that whish to implement this feature need to set property Mod.LoadSettingsCallback when they are started.

The value is a callback that receives an instance of ModSettings and ModSettingsChange, the latter being an holder for informations on the change event. Specifically, mods can use bool ModSettingsChange.HasChanged(string section, string key) to retrieve if a specific setting has changed.

For the time being, this feature is provided experimentally with a console command that opens a UI window.