Remove casting House4-5 back to House2
Now that fallback is available to any house type for local and remote, there is no need to force House4-5 back to House2. If quest is able to locate a House4-5 then great, otherwise it will fallback to one of the valid building types.
Added House1 back to validHouseTypes as this already existed in validBuildingTypes. Fallback and marker validation will help ensure a suitable house is found to host quest if one is available.