workaround for "water everywhere" texture issue on low-end hardware
It does not fix shadows issues or anything else, but still that's a
noticeable improvements for people that have to run DFU on that
hardware.

That looks like a very inexpensive workaround (most likely lost in the
noise), and it seems to work fine on my regular hardware, so maybe it
can be added directly into the standard shader?
Otherwise, say if the risk of regression is deemed too high, it could be
added to a variant of the shader, so it could be enabled at runtime; But
I don't know how to do such things...

Tested on Intel HD-3000 by Siva Machina on Discord;
Tested on nVidia GTX-960 by myself (for lack of regression)

Forums:
https://forums.dfworkshop.net/viewtopic.php?f=5&t=2519&p=30610#p30610