receive armor closes transient MessageBox
Use case: the player selects to EQUIP some item it cannot equip
(forbidden material, for example). EquipItem() opens a MessageBox to
tell it's not possible, chooseOne closes to top window which is the
MessageBox, and the inventory stays open.

Another implementation to set a flag, and close the inventory in
Update() when all transients have been closed by the player. The benefit
is that the player would see that the item couldn't be equipped.