Improve support for presets
Preset system required an update now that mod settings have been moved to PersistenDataPath.
I took the occasion to improve it and allow a mod to provide settings for another mod in a non-destructive way by bundling presets directly inside the mod.

Presets can be imported from the following locations:

- From the mod itself: 'presets.json'
- From other mods: '~/SourceModRoot/Assets/Presets/TargetModName/presetName.json'.
- From loose files: '~/StreamingAssets/Presets/TargetModName/presetName.json'