add automap shortcuts
Added automap shortcuts to setup

The most difficult part was replacing the fallback key feature: when the
key used to open the automap is used in an automap shortcut, it creates
a conflict because the automap implements toggle open (ie can be closed
with the same key used to open it). I reimplemented it like this:

- the key used to open the automap also closes the automap (toggle
open);
- shortcuts that used the same key are replaced with shortcuts using the
Home key, but with the same modifiers;
- if the key used to open the automap changes, just recreate all
shortcuts and tooltips.

Added Escape key to exit the automap for consistency and hardcoded
fallback

Added HotkeySequence.IsUp(With) methods used for toggle key

Added HotkeySequence.ToString() method

Removed hardcoded modifiers in tooltips texts

Renamed *Control- modifiers as *Ctrl- to shorter descriptions in automap
tooltips