Implement "clear_negativelegalrep" console command
Players are finding their legal reputations tanked for a variety of reasons. Often this is through their own actions, but certain quests can do this as well, leaving player effectively blacklisted from key regions and locked out of the main quest. This has proven to be frustrating for players and represents an ongoing burden on support channels.
This command provides a means to reset all negative legal rep and banishment state back to 0 so players can self-service their way of this mess if they don't wish to roleplay it.