Add IPlayerActivable interface for custom activations
GameObjects added by mods can use the interface IPlayerActivable to detect player interaction with the object.

There are no special requirements for the object or the purpose of this activation except that the interface must be implemented by a component and the object must have a collider for the activation to be detected.