Move save/new event unregistering into guild classes
This is so GuildManager.ClearMembershipData() does not call Leave() to unregister the events, because this method is for clearing memberships while loading etc (happens twice) and is not actually the player leaving the guilds. If mods add behaviour for leaving, it should not be initiated when just saving/loading/new game etc.

Now only RemoveMembership calls Leave().