Fix a bug when entering certain buildings riding a horse
Some extry doors are raised and cause issues when entered while on a horse since the controller height is larger so the raycast finds the streaming terrain and the player is then moved under the floor by PlayerEnterExit.SetStanding()
This changes that method to always use the standing controller height rather than the dynamic value.
Reproduce the bug by entering the palace at Issaqumbaa  in Abibon-Gora while riding. Before this fix you fall though the world.