Added click sound to new Quest Picker enhancement
Even though the quest picker wasn't in the original game, I added the click sound to picking a quest so that it's cohesive with how classic does it.

Also added the 'click sound' to some of the missing guild items that had a click sound on classic. Technically, you could make these sounds occur on the the pop-up and you'd only have to do this in one spot - but - you will get the sound playing twice in many circumstances that way.