Removed the continuous sound from scrolling
Classic does not have mouse scrolling in this window. Having the sound activate anytime you scroll, clips - and sounds terrible if you scroll too fast. So - I set it to where the traditional scroll sound plays when you click the arrows (like you had to do in classic) and just plays once like DGU-default when you use the scroll mouse. You could probably put a timer or something in between the sounds while scrolling so that it doesn't overlap - but I think that's just unnecessary. This replicates the classic sound as it is.