Make previous position of difficulty dagger slowly fade away
It can be difficult to assess by how much the difficulty dagger went up
or down after adding an advantage or a disadvantage, specially since
panels are being displayed on top of the dagger while it changes.

This adds a subtle animation that makes the previous position of the
dagger fade away for a second.