more strict options selection
The player must click on the alternatives he wants, instead of clicking
on any alternative to toggle it.

Classic mode had a benefit though, you could test the effect of an
alternative on cost and travel time without moving the mouse. So I moved
the toggling behavior to the mouse wheel.