Fix to ending drain effects
Sorry for earlier misinformation @ajrb. Had forgotten that DrainEffect is an especially sticky pestilence by design. To properly end the effect, it's necessary to "heal" the full amount of damage drained. Then effect will end by itself and bundle will be removed.
I've added a couple of extra helpers to assist with this. Tested with a character locally (will attach save to comment) and all stats are recovered now.