Set isPlayerInLocationRect before raising events
Mitigates an issue where if the target event handler throws an exception, it's possible to get caught in an endless exception loop.
Setting this flag before raising events also ensures that if target happens to check PlayerGPS.IsPlayerInLocationRect for some reason, the flag will have been updated prior to the event being called.